Moved new endpoint into PostsStatsService and updated endpoint
